requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.35
botocore==1.17.35
cffi==1.14.1
cryptography==3.0
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
psycopg2==2.8.5
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
